service-workers/service-worker/postmessage-msgport-to-client.https.html

<!DOCTYPE html>
<title>Service Worker: postMessage via MessagePort to Client</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
promise_test(async t => {
  const SCRIPT = 'resources/postmessage-msgport-to-client-worker.js';
  const SCOPE = 'resources/blank.html';

  const registration =
      await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
  await wait_for_state(t, registration.installing, 'activated');
  const frame = await with_iframe(SCOPE);

  t.add_cleanup(async() => {
    if (frame)
      frame.remove();
    if (registration)
      await registration.unregister();
  });

  let e = await new Promise(resolve => {
    const contentWindow = frame.contentWindow;
    contentWindow.navigator.serviceWorker.onmessage = resolve;
    contentWindow.navigator.serviceWorker.controller.postMessage('ping');
  });

  const port = e.ports[0];
  port.postMessage({value: 1});
  port.postMessage({value: 2});
  port.postMessage({done: true});
  e = await new Promise(resolve => { port.onmessage = resolve; });
  assert_equals(e.data.ack, 'Acking value: 1');
  e = await new Promise(resolve => { port.onmessage = resolve; });
  assert_equals(e.data.ack, 'Acking value: 2');
  e = await new Promise(resolve => { port.onmessage = resolve; });
  assert_true(e.data.done, 'done');

  }, 'postMessage MessagePorts from ServiceWorker to Client');
</script>
